+++ spine-cpp/include/spine/MathUtil.h
@@ -45,7 +45,7 @@ namespace spine {
 
 	public:
 		static const float Epsilon;
-		static const float Epsilon2;
+		static const float EpsilonSq;
 		static const float Pi;
 		static const float Pi_2;
 		static const float InvPi_2;